Fix package handling

+++ features/merchant/seller_mvc_creation/src/main/java/com/tokopedia/mvc/util/SharingUtil.kt
@@ -14,7 +14,7 @@ object SharingUtil {
         }
     }
 
-    fun shareToBroadCastChat(context: Context, voucherId: Int) {
+    fun shareToBroadCastChat(context: Context, voucherId: Long) {
         val broadCastChatUrl = "https://m.tokopedia.com/broadcast-chat/create/content?voucher_id=$voucherId"
         val broadCastChatIntent = BroadCastChatWebViewActivity.createNewIntent(
             context = context,
